+++ decomp/ST.exe/functions/00544940/decomp.c
@@ -3,7 +3,7 @@ void __fastcall FUN_00544940(AnonShape_00544940_E504B2A8 *param_1)
 
 {
   param_1->field_04DA = 0;
-  FUN_006b2800((int)PTR_008075a8,param_1->field_04D6,0,0);
+  FUN_006b2800(PTR_008075a8,param_1->field_04D6,0,0);
   FUN_006b3af0((int *)PTR_008075a8,param_1->field_04D6);
   return;
 }
